Properly implement NotesViewModel

See merge request fsc-clam/fsc-wheres-what!35

+++ app/src/main/java/io/gitlab/fsc_clam/fscwhereswhat/model/local/Note.kt
@@ -25,6 +25,6 @@ package io.gitlab.fsc_clam.fscwhereswhat.model.local
  */
 data class Note(
 	val comment: String,
-	val reference: Int,
+	val reference: Long,
 	val type: EntityType
 )
